Reuse existing voice channel categories
This allows us to remove the manage channels permission from the bot and avoid a scenario where a compromised bot deletes all the channels in the server.

Remove the strikes and voice channels tables, and the category class that aren't required anymore.
Remove the no longer required manage channels permission.

+++ src/events/client/removeExpiredVoiceChannels.ts
@@ -1,29 +1,20 @@
-import config from "../../config.js";
-import VoiceChannel from "../../tables/VoiceChannel.js";
-import {LessThanOrEqual} from "typeorm";
-import {DateTime} from "luxon";
-import {VoiceState} from "discord.js";
+import { VoiceBasedChannel, VoiceState } from "discord.js";
+import discord_server_config from "../../config/discord_server.js";
+
+function isBotManagedVoiceChannel(channel: VoiceBasedChannel): boolean {
+    const catName = channel.parent.name;
+    return catName.toLowerCase().startsWith(discord_server_config.hotdrop_vc_category_prefix.toLowerCase())
+        || catName.toLowerCase().startsWith(discord_server_config.strike_vc_category_prefix.toLowerCase());
+}
 
 // Dynamically delete pickdrop VCs as soon as everyone leaves
 export default {
     name: "voiceStateUpdate",
     once: false,
-    function: async (oldState: VoiceState, newState: VoiceState):Promise<null> => {
+    function: async (oldState: VoiceState, newState: VoiceState) => {
         const channel = oldState.channel || newState.channel;
-        if(!(channel.parent.id == config.vcCategory)) return;
-
-        const vc = await VoiceChannel.findOne({
-            where: {
-                channel: channel.id,
-                expires: LessThanOrEqual(DateTime.now().toMillis())
-            }
-        });
-
-        if (!vc) return;
-
-        if(channel && !channel.members.size) {
+        if (channel && isBotManagedVoiceChannel(channel) && channel.members.size == 0) {
             await channel.delete().catch((err) => console.log(err));
-            await vc.remove().catch((err) => console.log(err));
             console.log(`Expired & empty channel ${channel.id} deleted`);
         }
     }
